+++ R/utilities.R
@@ -132,8 +132,8 @@ toMissingOfType <- function(x, type) {
 #'
 #' @examples
 #' inputVector <- c(NA, 1, 5, 0, -1)
-#' log_safe(inputVector)
-log_safe <- function(x, base = exp(1), epsilon = ospsuiteUtilsEnv$LOG_SAFE_EPSILON) {
+#' logSafe(inputVector)
+logSafe <- function(x, base = exp(1), epsilon = ospsuiteUtilsEnv$LOG_SAFE_EPSILON) {
   x <- sapply(X = x, function(element) {
     element <- ospsuite.utils::toMissingOfType(element, type = "double")
     if (is.na(element)) {